fix: add dtype

+++ IMC/nn/sparse_metadata_encoder.py
@@ -212,8 +212,8 @@ class SparseMetadataEncoder(nn.Module):
         # Edge case: if no features are observed in the entire batch, return zero embedding
         if idxs.numel() == 0:
             if self.reduce:
-                return torch.zeros(B, self.out_dim, device=device)
-            return torch.zeros(B, S, self.out_dim, device=device)
+                return torch.zeros(B, self.out_dim, device=device, dtype=x.dtype)
+            return torch.zeros(B, S, self.out_dim, device=device, dtype=x.dtype)
 
         # Extract sample and feature indices for all observed values
         sample_idx = idxs[:, 0] # (N,), values in [0, B*S) indicating which sample
